Fix custom pin position issues on saving
Changes how chip descriptions are created, now accounting for custom pin layout (positions and presence of layout).

In addition, it fixes how the game sometimes falsly registers chips as unsaved when they have custom pin layouts.

It feels wrong to directly commit here.
