+++ Assets/Scripts/Graphics/World/CustomizationSceneDrawer.cs
@@ -6,6 +6,8 @@ using Seb.Types;
 using Seb.Vis;
 using UnityEngine;
 using System.Linq;
+using Seb.Vis.UI;
+using System.Collections.Generic;
 
 namespace DLS.Graphics
 {
@@ -22,8 +24,14 @@ namespace DLS.Graphics
 		static Vector2 mouseDownPos;
 		static Vector2 displayPosInitial;
 		static float displayScaleInitial;
+        public static PinInstance selectedPin;
+        public static bool isDraggingPin;
+        private static float lastValidOffset;
+        private static int lastValidFace;
+        public static bool isPinPositionValid;
+        static readonly float minPinSpacing = 0.025f;
 
-		static SubChipInstance CustomizeChip => ChipSaveMenu.ActiveCustomizeChip;
+        static SubChipInstance CustomizeChip => ChipSaveMenu.ActiveCustomizeChip;
 
 		public static bool IsPlacingDisplay => displayInteractState == DisplayInteractState.Placing;
 
@@ -41,6 +49,8 @@ namespace DLS.Graphics
 
 			bool chipResizeHascontrol = HandleChipResizing(chip);
 			HandleDisplaySelection(!chipResizeHascontrol);
+
+			HandlePinDragging();
 
 			if (SelectedDisplay == null)
 			{
@@ -379,13 +389,145 @@ namespace DLS.Graphics
 			}
 		}
 
-		public static void Reset()
+        static void HandlePinDragging()
+        {
+            if (!InteractionState.MouseIsOverUI)
+            {
+                // Start dragging a pin
+                if (InputHelper.IsMouseDownThisFrame(MouseButton.Left))
+                {
+                    if (InteractionState.ElementUnderMouse is PinInstance pin)
+                    {
+                        selectedPin = pin;
+                        isDraggingPin = true;
+                        lastValidOffset = pin.LocalPosY;
+                        lastValidFace = pin.face;
+                    }
+                }
+
+                if (isDraggingPin && selectedPin?.parent is SubChipInstance chip)
+                {
+                    Vector2 mouseWorld = InputHelper.MousePosWorld;
+                    Vector2 chipCenter = chip.Position;
+                    Vector2 localMouse = mouseWorld - chipCenter;
+                    Vector2 chipHalfSize = chip.Size / 2f;
+
+                    // Determine closest edge
+                    float distTop = Mathf.Abs(localMouse.y - chipHalfSize.y);
+                    float distBottom = Mathf.Abs(localMouse.y + chipHalfSize.y);
+                    float distRight = Mathf.Abs(localMouse.x - chipHalfSize.x);
+                    float distLeft = Mathf.Abs(localMouse.x + chipHalfSize.x);
+
+                    int closestFace = 0;
+                    float minDist = distTop;
+
+                    if (distRight < minDist) { closestFace = 1; minDist = distRight; }
+                    if (distBottom < minDist) { closestFace = 2; minDist = distBottom; }
+                    if (distLeft < minDist) { closestFace = 3; }
+
+                    selectedPin.face = closestFace;
+
+                    float pinHeight = SubChipInstance.PinHeightFromBitCount(selectedPin.bitCount);
+
+                    float maxOffset;
+                    float offsetAlongFace;
+
+                    bool shouldSnapToGrid = Project.ActiveProject.ShouldSnapToGrid;
+
+                    if (closestFace == 0 || closestFace == 2)
+                    {
+                        // Horizontal face - move along X axis
+                        maxOffset = chipHalfSize.x - pinHeight / 2f;
+                        offsetAlongFace = shouldSnapToGrid ? GridHelper.ClampToGrid(localMouse.x, -maxOffset, maxOffset) :
+                            Mathf.Clamp(localMouse.x, -maxOffset, maxOffset);
+                    }
+                    else
+                    {
+                        // Vertical face - move along Y axis
+                        maxOffset = chipHalfSize.y - pinHeight / 2f;
+                        offsetAlongFace = shouldSnapToGrid ? GridHelper.ClampToGrid(localMouse.y, -maxOffset, maxOffset) :
+                            Mathf.Clamp(localMouse.y, -maxOffset, maxOffset);
+                    }
+
+                    selectedPin.LocalPosY = offsetAlongFace;
+
+                    PinInstance overlappedPin;
+                    isPinPositionValid = !DoesPinOverlap(selectedPin, out overlappedPin);
+
+                    // End drag on mouse release
+                    if (InputHelper.IsMouseUpThisFrame(MouseButton.Left))
+                    {
+                        if (isPinPositionValid)
+                        {
+                            lastValidOffset = offsetAlongFace;
+                            lastValidFace = selectedPin.face;
+
+                            if (!ChipCustomizationMenu.isCustomLayout)
+                            {
+                                ChipCustomizationMenu.isCustomLayout = true;
+                                UI.GetWheelSelectorState(ChipCustomizationMenu.ID_LayoutOptions).index = 1;
+                                ChipSaveMenu.ActiveCustomizeChip.SetCustomLayout(true);
+                            }
+                        }
+                        else
+                        {
+                            selectedPin.LocalPosY = lastValidOffset;
+                            selectedPin.face = lastValidFace;
+                        }
+
+                        isDraggingPin = false;
+                        selectedPin = null;
+                        isPinPositionValid = true;
+                    }
+                }
+            }
+        }
+
+        public static void Reset()
 		{
 			SelectedDisplay = null;
 			displayInteractState = DisplayInteractState.None;
 		}
 
-		enum DisplayInteractState
+        public static bool DoesPinOverlap(PinInstance pin, out PinInstance overlappedPin)
+        {
+            overlappedPin = null;
+            if (!(pin.parent is SubChipInstance chip)) return false;
+
+            // Get all pins on the same chip to check pins on the same face as selectedpin
+            List<PinInstance> pinsToCheck = new List<PinInstance>();
+            pinsToCheck.AddRange(chip.InputPins);
+            pinsToCheck.AddRange(chip.OutputPins);
+
+            foreach (PinInstance otherPin in pinsToCheck)
+            {
+                if (otherPin == pin) continue;
+
+                // Only check pins on the same face
+                if (otherPin.face != pin.face) continue;
+
+                float distanceAlongFace = Mathf.Abs(pin.LocalPosY - otherPin.LocalPosY);
+
+                // Calculate minimum required spacing based on pin sizes
+                float pinHeight = SubChipInstance.PinHeightFromBitCount(pin.bitCount);
+
+                float otherPinHeight = SubChipInstance.PinHeightFromBitCount(otherPin.bitCount);
+
+                // Required space is half each pin's height plus some buffer
+                float requiredSpacing = (pinHeight + otherPinHeight) / 2f + minPinSpacing;
+
+                if (distanceAlongFace < requiredSpacing)
+                {
+                    overlappedPin = otherPin;
+                    return true;
+                }
+            }
+
+            return false;
+        }
+
+
+        enum DisplayInteractState
 		{
 			None,
 			Moving,
